Just wanted to show my changes. DO NOT MERGE. I need a better sprite sheet. They hop everywhere. It works though. - Deleted SpriteClasses - Added Singular ItemSprite - Modified Game1 IItemSprite List - Deleted .png files - Added .png files
This was the route I wanted to take originally, but since he stated to use classes for all, I was like uh okay. Yeah that was a mistake :). I imagine Ill be implementing classes still but those will be object classes not sprite classes. AKA for how the impact link and such.